+++ res/changelog.md
@@ -9,6 +9,13 @@
  - Purge unused files and dependencies.
 
 ## Changelog
+### 2022-10-26
+ - Re-added ConfirmationBox
+ - Added InputBox
+ - Added context menu item for setting folder titles
+ - BUGFIX: WordStatsManager.extendStats is run on WSFiles with no stats, resulting in an exception.
+ 
+
 ### 2022-10-25
  - Added context menu items to set word statistics recording state on folders, and commands to set it for the focused file's parent.
  - BUGFIX: ProgressView was not monitoring changes to statistics recording state.